+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2018-01-29
+### Added
+- `Router` scoped common middlewares.
+- Tests with common middlewares.
+
+### Fixed
+- Filter interfaces before pushing to the radix tree.
+
+### Removed
+- Global scoped common middlewares.
+
 ## [0.8.0] - 2018-01-29
 ### Added
 - Variable to hold and set common middleware between endpoints.
@@ -93,6 +104,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
+[0.9.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/gbrlsnchs/httpmux/compare/v0.6.0...v0.6.1